Bad use of quint8
For whatever reason I had decided to use a quint8 type for some of the
parsing which in effect was causing bad parsing of some of the tags
since I should have probably been using quint32.  That is fixed, however
there does seem to be an issue with overrun -- sometimes the parser
comes close to the end of the tag and there's a bad frame size input.
Not sure why yet.

+++ id3tagparser.h
@@ -28,7 +28,7 @@ private:
     void getTagID3v2(QFile& file, Tag* tag) const;
 
     QString parseFrameData(char* buf, const quint8 size) const;
-    quint8 getFrameDataSize(char* buf, char id3ver) const;
+    unsigned int getFrameDataSize(char* buf, char id3ver) const;
     unsigned char parseTrack(const QString& str) const;
 
 };
